requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
autopep8==2.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
distlib==0.3.6
Django==4.2.1
django-3-jet==1.0.8
django-4-jet @ https://github.com/Barukimang/django-jet/archive/dev.zip#sha256=73bfdc1ee258ae78c2a1149bea889de8d8a70b1b5807c769c5277f1b40bab499
django-cleanup==7.0.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-jet==1.0.8
django-paypal==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
filelock==3.12.0
freetype-py==2.3.0
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.5.0
pipenv==2023.4.20
platformdirs==3.2.0
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==4.0.0
requests==2.30.0
rlPyCairo==0.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
sweetify==2.3.1
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==2.0.2
virtualenv==20.22.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xhtml2pdf==0.2.10
